requirements.txt
flask
google-cloud-firestore
pytest
functions-framework
